+++ lib/bluepan_client/responses/base_response.rb
@@ -9,6 +9,7 @@ module BluepanClient
       default: :default_parsed_body,
     })
     attribute :success, Boolean, lazy: true, default: :default_success
+    attribute :error_message, String, lazy: true, default: :default_error_message
 
     private
 
@@ -24,6 +25,14 @@ module BluepanClient
 
     def default_parsed_body
       JSON.parse(body)
+    rescue JSON::ParserError
+      nil
+    end
+
+    def default_error_message
+      return nil if success?
+      return parsed_body[:error] if parsed_body.present? && parsed_body[:error]
+      "Status: #{raw_response.code}; Error message not found in `#{body}`"
     end
 
   end